+++ tests/testthat/test-get_code.R
@@ -637,7 +637,7 @@ testthat::test_that("understands @ usage and do not treat rhs of @ as objects (o
 testthat::test_that("library() and require() are always returned", {
   code <- c(
     "set.seed(1)",
-    "library(scda)",
+    "library(random.cdisc.data)",
     "require(dplyr)",
     "library(MultiAssayExperiment)",
     "x <- 5",
@@ -648,7 +648,7 @@ testthat::test_that("library() and require() are always returned", {
     get_code(tdata, datanames = "x"),
     paste(
       warning_message,
-      "library(scda)",
+      "library(random.cdisc.data)",
       "require(dplyr)",
       "library(MultiAssayExperiment)",
       "x <- 5",
@@ -663,7 +663,7 @@ testthat::test_that("library() and require() are always returned", {
 testthat::test_that("data() call is returned when data name is provided as is", {
   code <- c(
     "set.seed(1)",
-    "library(scda)",
+    "library(random.cdisc.data)",
     "require(dplyr)",
     "library(MultiAssayExperiment)",
     "data(miniACC, envir = environment())",
@@ -674,7 +674,7 @@ testthat::test_that("data() call is returned when data name is provided as is", 
     get_code(tdata, datanames = "x"),
     paste(
       warning_message,
-      "library(scda)",
+      "library(random.cdisc.data)",
       "require(dplyr)",
       "library(MultiAssayExperiment)",
       "data(miniACC, envir = environment())",
@@ -687,7 +687,7 @@ testthat::test_that("data() call is returned when data name is provided as is", 
 testthat::test_that("data() call is returned when data name is provided as a character", {
   code <- c(
     "set.seed(1)",
-    "library(scda)",
+    "library(random.cdisc.data)",
     "require(dplyr)",
     "library(MultiAssayExperiment)",
     "data('mtcars')",
@@ -698,7 +698,7 @@ testthat::test_that("data() call is returned when data name is provided as a cha
     get_code(tdata, datanames = "z"),
     paste(
       warning_message,
-      "library(scda)",
+      "library(random.cdisc.data)",
       "require(dplyr)",
       "library(MultiAssayExperiment)",
       "data(\"mtcars\")",